# unexpand: improve performances (1.50 faster than GNU)