```
hyperfine --warmup 3 "/usr/bin/unexpand moby64.txt"                     "./target/release/coreutils.prev unexpand moby64.txt" "./target/release/coreutils unexpand moby64.txt"
Benchmark 1: /usr/bin/unexpand moby64.txt
  Time (mean ± σ):      95.8 ms ±   3.8 ms    [User: 81.1 ms, System: 14.6 ms]
  Range (min … max):    89.6 ms … 104.8 ms    31 runs

Benchmark 2: ./target/release/coreutils.prev unexpand moby64.txt
  Time (mean ± σ):     952.3 ms ±   3.0 ms    [User: 846.4 ms, System: 105.6 ms]
  Range (min … max):   949.4 ms … 958.1 ms    10 runs

Benchmark 3: ./target/release/coreutils unexpand moby64.txt
  Time (mean ± σ):      63.7 ms ±   2.2 ms    [User: 52.9 ms, System: 10.7 ms]
  Range (min … max):    60.8 ms …  70.2 ms    46 runs

Summary
  ./target/release/coreutils unexpand moby64.txt ran
    1.50 ± 0.08 times faster than /usr/bin/unexpand moby64.txt
   14.95 ± 0.53 times faster than ./target/release/coreutils.prev unexpand moby64.txt
```